# openapi.json
{
  "components": {
    "schemas": {
...
      "Catalog2": {
        "description": "Tipo de Moneda - Catálogo N.° 02",
        "enum": [
          "PEN",
          "USD",
          "EUR"
        ],
        "title": "Catalog2",
        "type": "string"
      },
...
      "CreditNote": {
        "description": "Nota de Crédito Electrónica - Tipo 07.\n\nRS N° 300-2014/SUNAT, Anexo 1:\n- Serie: debe iniciar con B/C o F/E según tipo de documento afectado\n- Debe referenciar el comprobante afectado",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "comprobanteAfectadoSerieNumero": {
            "title": "Comprobanteafectadoserienumero",
            "type": "string"
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/DocumentoVentaDetalle"
            },
            "title": "Detalles",
            "type": "array"
          },
          "fechaEmision": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fechaemision"
          },
          "igvTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Igvtotal"
          },
          "importeTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importetotal"
          },
          "moneda": {
            "$ref": "#/components/schemas/Catalog2",
            "default": "PEN"
          },
          "numero": {
            "minimum": 1.0,
            "title": "Numero",
            "type": "integer"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "serie": {
            "description": "Serie de nota de crédito",
            "pattern": "^[BCbc][A-Za-z0-9]{2,3}$",
            "title": "Serie",
            "type": "string"
          },
          "sustentoDescripcion": {
            "title": "Sustentodescripcion",
            "type": "string"
          },
          "valorVentaTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valorventatotal"
          }
        },
        "required": [
          "serie",
          "numero",
          "comprobanteAfectadoSerieNumero",
          "sustentoDescripcion",
          "proveedor",
          "cliente",
          "detalles"
        ],
        "title": "CreditNote",
        "type": "object"
      },
      "DebitNote": {
        "description": "Nota de Débito Electrónica - Tipo 08.\n\nRS N° 300-2014/SUNAT, Anexo 1:\n- Serie: debe iniciar con B/C o F/E según tipo de documento afectado\n- Debe referenciar el comprobante afectado",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "comprobanteAfectadoSerieNumero": {
            "title": "Comprobanteafectadoserienumero",
            "type": "string"
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/DocumentoVentaDetalle"
            },
            "title": "Detalles",
            "type": "array"
          },
          "fechaEmision": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fechaemision"
          },
          "igvTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Igvtotal"
          },
          "importeTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importetotal"
          },
          "moneda": {
            "$ref": "#/components/schemas/Catalog2",
            "default": "PEN"
          },
          "numero": {
            "minimum": 1.0,
            "title": "Numero",
            "type": "integer"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "serie": {
            "description": "Serie de nota de débito",
            "pattern": "^[BCbc][A-Za-z0-9]{2,3}$",
            "title": "Serie",
            "type": "string"
          },
          "sustentoDescripcion": {
            "title": "Sustentodescripcion",
            "type": "string"
          },
          "valorVentaTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valorventatotal"
          }
        },
        "required": [
          "serie",
          "numero",
          "comprobanteAfectadoSerieNumero",
          "sustentoDescripcion",
          "proveedor",
          "cliente",
          "detalles"
        ],
        "title": "DebitNote",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Invoice": {
        "description": "Factura Electrónica - Tipo 01.\n\nRS N° 300-2014/SUNAT, Anexo 1:\n- Serie: debe iniciar con F (factura) o B (boleta)\n- Número: correlativo\n- Tipo de operación: Catálogo N.° 51, default 0101 (Venta interna)",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "detalles": {
            "items": {
              "$ref": "#/components/schemas/DocumentoVentaDetalle"
            },
            "title": "Detalles",
            "type": "array"
          },
          "fechaEmision": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fechaemision"
          },
          "igvTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Igvtotal"
          },
          "importeTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importetotal"
          },
          "moneda": {
            "$ref": "#/components/schemas/Catalog2",
            "default": "PEN"
          },
          "numero": {
            "minimum": 1.0,
            "title": "Numero",
            "type": "integer"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "serie": {
            "description": "Serie de factura (F001) o boleta (B001)",
            "pattern": "^[FBfb][A-Za-z0-9]{2,3}$",
            "title": "Serie",
            "type": "string"
          },
          "tipoOperacion": {
            "$ref": "#/components/schemas/Catalog51",
            "default": "0101"
          },
          "valorVentaTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valorventatotal"
          }
        },
        "required": [
          "serie",
          "numero",
          "proveedor",
          "cliente",
          "detalles"
        ],
        "title": "Invoice",
        "type": "object"
      },
...
      "Perception": {
        "description": "Comprobante de Percepción Electrónico - Tipo 40.\n\nRS N° 274-2015/SUNAT, Anexo 1:\n- Serie: P### (P001, P002, etc.)\n- Régimen: Catálogo N.° 22",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "fechaEmision": {
            "format": "date",
            "title": "Fechaemision",
            "type": "string"
          },
          "importeTotalCobrado": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Importetotalcobrado"
          },
          "importeTotalPercibido": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Importetotalpercibido"
          },
          "numero": {
            "minimum": 1.0,
            "title": "Numero",
            "type": "integer"
          },
          "operaciones": {
            "items": {
              "$ref": "#/components/schemas/PercepcionRetencionOperacion"
            },
            "title": "Operaciones",
            "type": "array"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "serie": {
            "description": "Serie de percepción (P###)",
            "pattern": "^P\\d{3}$",
            "title": "Serie",
            "type": "string"
          },
          "tipoRegimen": {
            "$ref": "#/components/schemas/Catalog22"
          },
          "tipoRegimenPorcentaje": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Tiporegimenporcentaje"
          }
        },
        "required": [
          "serie",
          "numero",
          "fechaEmision",
          "proveedor",
          "cliente",
          "importeTotalPercibido",
          "importeTotalCobrado",
          "tipoRegimen",
          "tipoRegimenPorcentaje",
          "operaciones"
        ],
        "title": "Perception",
        "type": "object"
      },
...
      "Retention": {
        "description": "Comprobante de Retención Electrónico - Tipo 20.\n\nRS N° 274-2015/SUNAT, Anexo 1:\n- Serie: R### (R001, R002, etc.)\n- Régimen: Catálogo N.° 23",
        "properties": {
          "importeTotalPagado": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Importetotalpagado"
          },
          "importeTotalRetenido": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Importetotalretenido"
          },
          "operaciones": {
            "items": {
              "$ref": "#/components/schemas/PercepcionRetencionOperacion"
            },
            "title": "Operaciones",
            "type": "array"
          },
          "serie": {
            "description": "Serie de retención (R###)",
            "pattern": "^R\\d{3}$",
            "title": "Serie",
            "type": "string"
          },
          "tipoRegimen": {
            "$ref": "#/components/schemas/Catalog23"
          },
          "tipoRegimenPorcentaje": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Tiporegimenporcentaje"
          }
        },
        "required": [
          "serie",
          "importeTotalRetenido",
          "importeTotalPagado",
          "tipoRegimen",
          "tipoRegimenPorcentaje",
          "operaciones"
        ],
        "title": "Retention",
        "type": "object"
      },
      "SignedXmlResponse": {
        "description": "Response containing signed XML document.",
        "properties": {
          "signed_xml": {
            "title": "Signed Xml",
            "type": "string"
          }
        },
        "required": [
          "signed_xml"
        ],
        "title": "SignedXmlResponse",
        "type": "object"
      },
      "SummaryDocuments": {
        "description": "Resumen Diario - RC.\n\nRS N° 300-2014/SUNAT, Anexo 1:\n- Identificador: RC-YYYYMMDD-NNNN\n- Fecha de emisión de los comprobantes resumidos",
        "properties": {
          "comprobantes": {
            "items": {
              "$ref": "#/components/schemas/SummaryDocumentsItem"
            },
            "title": "Comprobantes",
            "type": "array"
          },
          "fechaEmisionComprobantes": {
            "format": "date",
            "title": "Fechaemisioncomprobantes",
            "type": "string"
          },
          "numero": {
            "title": "Numero",
            "type": "integer"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          }
        },
        "required": [
          "numero",
          "fechaEmisionComprobantes",
          "proveedor",
          "comprobantes"
        ],
        "title": "SummaryDocuments",
        "type": "object"
      },
...
      "XmlResponse": {
        "description": "Response containing generated XML document.",
        "properties": {
          "xml": {
            "title": "Xml",
            "type": "string"
          }
        },
        "required": [
          "xml"
        ],
        "title": "XmlResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "openUBL",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/credit-note/create": {
      "post": {
        "description": "Generate a CreditNote XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createCreditNote",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditNote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Credit Note"
      }
    },
    "/api/v1/debit-note/create": {
      "post": {
        "description": "Generate a DebitNote XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createDebitNote",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebitNote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Debit Note"
      }
    },
    "/api/v1/invoice/create": {
      "post": {
        "description": "Generate an Invoice XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createInvoice",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invoice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invoice"
      }
    },
    "/api/v1/perception/create": {
      "post": {
        "description": "Generate a Perception XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createPerception",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Perception"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Perception"
      }
    },
    "/api/v1/retention/create": {
      "post": {
        "description": "Generate a Retention XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createRetention",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Retention"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Retention"
      }
    },
    "/api/v1/sign": {
      "post": {
        "description": "Sign an arbitrary UBL XML document with a PEM certificate/key pair or a PFX/P12 container.\n\nRequired body fields (choose one credential mode):\n    * PEM mode: `cert_pem` and `key_pem`.\n    * PFX/P12 mode: `pfx_base64` (standard base64 of the file) and `pfx_password`.\n\nOptional body fields:\n    * `xml`: The XML string to sign.\n    * `signature_id`: The signature ID (defaults to `SignSUNAT`).\n\nReturns:\n    200: `{\"signed_xml\": \"...\"}` with the signed XML.",
        "operationId": "signXml",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedXmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign Xml"
      }
    },
    "/api/v1/summary-documents/create": {
      "post": {
        "description": "Generate a SummaryDocuments XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createSummaryDocuments",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryDocuments"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Summary Documents"
      }
    },
...
    "/api/v1/voided-documents/create": {
      "post": {
        "description": "Generate a VoidedDocuments XML document.\n\nThe `validate` query parameter controls SUNAT validation and defaults to `true`.\n\nReturns:\n    200: `{\"xml\": \"...\"}` with the generated XML.\n    422: Validation failed or the request body is invalid.",
        "operationId": "createVoidedDocuments",
        "parameters": [
          {
            "in": "query",
            "name": "validate",
            "required": false,
            "schema": {
              "default": true,
              "title": "Validate",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidedDocuments"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XmlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Voided Documents"
      }
    }
  }
}